Added backend routes for all pages
Now user creation, login and logout should work. These were implemented using the new routes, so that the backend knows if the user has logged in. This way it's possible to e.g. have only one main menu template, since the backend can tell the template if the user has a save file, and the template can then disable the "Load game" button if that's not the case. The backend can now also handle redirecting the user to the right page e.g. after login.

Also refactored the file structure, so that e.g. all images are in one directory. The files needed to be moved, since Flask needs .html and static files to be in "templates" and "static" directories.